mv, lenses to optics folder
