+++ 6_hybridlens_design.py
@@ -15,7 +15,7 @@ import yaml
 from torchvision.utils import save_image
 from tqdm import tqdm
 
-from deeplens.hybridlens import HybridLens
+from deeplens.optics import HybridLens
 from deeplens.optics.loss import PSFLoss
 from deeplens.utils import set_logger, set_seed
