+++ deeplens/optics/__init__.py
@@ -88,3 +88,11 @@ class DeepObj:
 
 from .material import *
 from .light import *
+
+# Lens classes
+from .lens import Lens
+from .geolens import GeoLens
+from .hybridlens import HybridLens
+from .diffraclens import DiffractiveLens
+from .paraxiallens import ParaxialLens
+from .psfnetlens import PSFNetLens
