+++ deeplens/optics/paraxiallens.py
@@ -14,9 +14,9 @@ Reference:
 import numpy as np
 import torch
 
-from deeplens.lens import Lens
-from deeplens.optics.config import DEPTH, EPSILON, PSF_KS
-from deeplens.optics.imgsim import conv_psf_depth_interp, conv_psf_occlusion
+from .lens import Lens
+from .config import DEPTH, EPSILON, PSF_KS
+from .imgsim import conv_psf_depth_interp, conv_psf_occlusion
 
 
 class ParaxialLens(Lens):
