+++ docs/quickstart.rst
@@ -129,7 +129,7 @@ For hybrid refractive-diffractive lens systems:
 .. code-block:: python
 
     import torch
-    from deeplens.hybridlens import HybridLens
+    from deeplens.optics import HybridLens
     
     # Set double precision for accurate wave optics
     torch.set_default_dtype(torch.float64)
@@ -147,7 +147,7 @@ For simple paraxial lens model with Circle of Confusion (CoC):
 
 .. code-block:: python
 
-    from deeplens.paraxiallens import ParaxialLens
+    from deeplens.optics import ParaxialLens
     
     lens = ParaxialLens(
         foclen=50.0,           # Focal length in mm
@@ -167,7 +167,7 @@ For pure diffractive optical elements using wave propagation:
 
 .. code-block:: python
 
-    from deeplens.diffraclens import DiffractiveLens
+    from deeplens.optics import DiffractiveLens
     
     # Load from file
     lens = DiffractiveLens(
