+++ test/test_paraxiallens.py
@@ -5,7 +5,7 @@ Tests for deeplens/paraxiallens.py - Paraxial lens model.
 import pytest
 import torch
 
-from deeplens.paraxiallens import ParaxialLens
+from deeplens.optics import ParaxialLens
 from deeplens.optics.config import DEPTH
